pyproject.toml
[project]
name = "ax-agent-studio"
version = "0.1.0"
description = "Monitor and manage aX platform MCP agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "psutil>=5.9.0",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.9",
    "openai>=1.0.0",
    "langgraph>=0.1.0",
    "pyyaml>=6.0.0",
    "langchain-aws>=1.0.0",
    "langchain-google-genai>=3.0.0",
    "langchain-anthropic>=1.0.0",
    "langchain-openai>=1.0.1",
    "langchain-mcp-adapters>=0.1.11",
]
